fz.tint_scene and tz.tint_scene
